Fix integer overflow in mutt_convert_string().
In the case of a *very* large message header (which is possible via a
compressed encrypted message) it's possible to overflow the incorrect
assignment of strlen() to an int local variable.

Thanks to Tavis Ormandy for the bug report and patch, which this
commit is based upon.

Although Tavis wasn't able to find an exploit, it was almost possible
to make ob small and obl big, which would have allowed attacker
control of a heap corruption.

Change the strlen() to assign directly to ibl (of type size_t).  This
prevents signed to unsigned conversion of len to ibl, which make the
attack almost possible.

Note that ibl should reflect the number of bytes to be converted by
iconv(), so the change of ibl to strlen() instead of (strlen() + 1) is
intentional, and correct.

ob is allocated with an additional byte for a trailing nul, which is
appended after the conversion.