When expanding local paths, normalize to remove a trailing '/'.
Commit 986e9e74 normalized Maildir/mh paths upon opening, to aid in
mailbox comparison issues that can crop up (as the IMAP path
normalization does).

Unfortunately, this caused other problems in cases where users were
explicitly adding a trailing '/' to their config settings, such as
with buffy mailboxes, or with $spoolfile.

To normalize properly, we need to do as the IMAP code does - both on
the context opening, and expand_path (as imap_expand_path() and
imap_fix_path() do).

This also helps other cases; for example, the file browser returns
entries without a trailing '/', while tab completion in the editor
menu appends a trailing '/'.

Right now the only regression I'm aware of is with $header_cache on
first use, which I will document in the next commit.  If it turns out
there are more important regressions I may have to back this and the
original commit out.