Document $header_cache behavior change wrt directories.
Although not documented, Mutt would previously intepret a
$header_cache value ending in '/' as a directory even if it didn't
exist.  The new DT_PATH normalization prevents this possibility, so a
directory will need to be created in advance.

I'm not fond of "fixing" regressions with documentation, but I believe
this is a small issue, easily worked around for first-time use, and
the benefits of the normalization are worth this small change.