Set gsasl hostname callback value.
This is needed for GSSAPI, and apparently DIGEST-MD5 too.

The gsasl documentation is a little confusing, saying it "should be
the local host name of the machine", however the imap/auth_gss.c code
seems to be using the server name, and the msmtp source also uses the
server name for this callback.

Thanks to brian m. carlson and Gábor Gombás for reporting this issue
in Debian ticket 1010915, and an additional thanks to brian for
quickly testing the fix.