Document $sendmail invocation behavior.
This variable is handled differently from other "command" variables in
Mutt.  It's tokenized by space and then executed via execvp().  This
means spaces in command/arguments are not supported, and neither is
shell quoting.

I don't know if it was done this way out of some security concern, but
it seems like using mutt_system() and mutt_buffer_quote_filename() for
recipient arguments should at least be investigated.