Fix browser completion path expansion to preserve a trailing slash.
The browser lists the contents of a directory passed as 'f' when it
has a trailing slash; without, it lists everything matching that name
in the parent directory.

Since the browser does its own relative path expansion, we can just
use mutt_buffer_path_norel() to keep a trailing slash in 'f'.