Fix header folding my_width() calculation.
After calculating the width of a character, the routine would only
increment the string pointer by one byte.  Any errors returned by
mbtowc() would also increment the width by one.  This means multibyte
characters would overcount width by the number of bytes minus one.

Change it to check the return value and use that value to increment
the string pointer.

Change mbtowc() to look at the whole rest of the string instead of
just MB_CUR_MAX, as the manpage says even MB_CUR_MAX may not be enough
in some circumstances.

Since we calculate strlen, use that as well as '\0' for the loop
termination check.  Also check for mbtowc() returning 0 just for extra
safety.

Reset the internal mbstate_t before converting, and on any error.

If mbtowc() returns an error, use replacment_char() as a substitue for
width calcluation, as mutt_strwidth() and other parts of Mutt do.