Fix $pgp_sort_keys sorting.
Both gpgme and pgpkey used nonsensical comparison return values, for
example: "return r > 0".

Adjust numeric comparisons to use mutt_numeric_cmp() and have the
comparator return the result of the actual comparison.

Adjust the "trust" sorting of gpgme to be the same as classic-pgp:
putting restrictions at the bottom, but reverse sorting validity,
length and timestamp values so they come first in the list.