Improve mbox parsing to address ticket 415.
Check for newline separators between messages.  To allow parsing of
old mbox files, merely warn in the debug log, but continue allowing
parsing.

Track when a content-length delimited message is not follow by a From_
line.  In that case, abort since it means there is some kind of
corruption and a message will otherwise likely be silently ignored.