Fix mutt_read_rfc822_line() to use is_email_wsp().
ISSPACE() uses isspace() which is locale-dependent.  On some
platforms, unexpected 8-bit chars, such as 0xa0 or 0x85 return true.

When using $edit_headers, this can result in Subject: lines being
truncated if a multi-byte character ending in one of these values is
at the end of a line.

There are probably other bugs that could be triggered by this, such as
in IMAP parsing.  However, I need more time to investigate before
making large-scale changes that could introduce new bugs.