Move SigWinch handling before refresh in menus.
This will prevent an unneeded double-refresh after an endwin() when we
set SigWinch in the next commit.

In the pager, change RETWINCH handling to account for a sigwinch while
in the handler.  Previously a redraw would occur and use/free the
Resize data, but since the SigWinch check now occurs before the
redraw, keep existing data.