Abort imap_fast_trash() if previously checkpointed.
We don't want to copy the deleted flag over to the trash folder too.

I looked into various ways to keep the UID COPY, but they lead to
niggling issues with error handling along with handling if the server
sends flag updates back to the client.

So for that (hopefully rare) case, abort the fast trash and just use a
regular copy.