Reset header color after mutt_set_flag().
I partially changed this to lazily update after a thread update in
commit c9fa0414, but unfortunately didn't investigate the reason for
the color update while setting a flag.  Since it was that way, I
assumed it was for a purpose.

However, it turns out there is no need to actively set the header
color in that function.  Many places in Mutt already simply reset the
color values to 0 to invalidate and cause a recheck later.

Setting the color there so can even be detrimental, if the user has
slow 'color index' lines.  For example doing a <tag-pattern>~A will
cause the color to be computed for the *entire* mailbox.  Now, the
user ought to not have a slow color index line, but if they do, this
causes unnecessary pain.

Note that the header->color doesn't have an actual "unset" value,
which could also help performance.  Maybe in the future in master
branch.