Fix <collapse-all> behavior for sort=reverse-threads.
When uncollapsing, _mutt_traverse_thread() returns the virtual number
of the root message in the thread.  <collapse-thread> directly sets
menu->current to this value to cause the cursor to be on the *first*
message of the thread (which isn't the same as the root message when
sort=reverse-threads).  <collapse-all> finds the corresponding message
by searching for it after re-indexing.

However, when collapsing, _mutt_traverse_thread() had code to try and
find the *first* message in the thread and return that virtual number.
<collapse-thread> then did the same trick, because the old first
message is now the new root message for sort=reverse-threads.

However, that cleverness caused a bug for <collapse-all> - it can't
use that virtual number directly, and it can't "find" the message at
the index after reindexing.

To fix this, remove the cleverness from _mutt_traverse_thread() when
collapsing.  Return the virtual number of the root.  Add searching
behavior for <collapse-thread> to fix its behavior.