Fix GPGME build failure on MacOS.
Commit 012981e8 (in release 2.2.9) updated the GPGME autoconf files,
to fix a build issue with newer GPGME releases.

Unfortunatley that caused a build issue for hosts where the gpg-error
header files aren't in the include path.  The newer autoconf file
expect GPG_ERROR_CFLAGS to be added to the list of flags for the
compiler.

Thanks to Will Yardley for reporting the issue and quickly testing the
proposed fix.