configure.ac: Fix ICONV_NONTRANS probe
The standard iconv function uses char ** even for its input argument.
With a const char ** argument, ICONV_NONTRANS is incorrectly set to 1
if the compiler produces an error for such incompatible pointer types.

Although as far as I can see, the only thing that accomplishes is
disabling an assert, so it probably does not matter much.