Fix configure.ac AM_ICONV result checking.
--disable-iconv seeds the cache value variable, $am_cv_func_iconv, to
"no" to skip the test and disable iconv.

However, the result of an existing and *working* AM_ICONV test is
stored in $am_func_iconv.  The call to AC_DEFINE(HAVE_ICONV) depends
on that variable.

Mutt was improperly checking $am_cv_func_iconv below.

In the event of an existing but broken iconv, this would result in
HAVE_ICONV being undefined, but mutt_idna.o being added to
MUTT_LIB_OBJECTS.  This would cause both the stub functions in
mutt_idna.h and the full functions in mutt_idna.c to be defined,
giving a compilation error.

Thanks to @juanitotc for the bug report and helping to test the fix.