Remove "." from MimeSpecials (RFC2045 tspecials).
RFC2045 declares tspecials to not include a period.  So encoding it
inside 2231 parameter values, or adding double quotes to MIME
parameter values (such as in Content-Type and Content-Disposition)
because of the presence of a period, while not illegal, is
unnecessary.

Ticket 491 reported a case where some Android clients were not
decoding the "." in an attachment filename extension for some unknown
reason.  While this is clearly a bug in the Android client, it's also
true that the period need not have been encoded in the first place.

However, "." couldn't simply be removed from MIMESpecials, because it
was also used for RFC2047 encoding.  RFC2047 encoded-words are used in
non-MIME headers, and cannot be double quoted.

Create a third "specials" list, RFC2047Specials, adding back in the
".", to keep RFC2047 encoding the same as it was.  Add a comment as to
why it exists, to prevent someone from making the unfortunate mistake
of thinking it can just be changed back to use MIMESpecials.

Thanks to Peter Seiderer for reporting the issue and suggesting a fix,
which this patch is based upon.