Fix isdigit() argument to unsigned char in pager is_ansi() call.
In a refactor, in commit 0ac6b609cd9e4588f00221b5c66e0fd5f584da88 a
few years ago, I inadvertantly changed the type of the is_ansi()
parameter, and thus changed the isdigit() argument to char.

This is not allowed, and needs to be cast to (unsigned char) for
correct behavior on some platforms.

Thanks to Thomas Klausner (@_wiz_) for pointing out the bug and for
the patch.