Add documentation about <buffy-cycle> and its keybinding.
Add a subsection to the "Editing Input Fields" section of the manual,
explaining about the <buffy-cycle> function, and when it is used.

Also, add an explanation for a previously undocumented behavior: in
other filename and mailbox prompts, typing the <buffy-cycle> key
binding falls back and invokes the <complete> function.

Mention the <quote-char> function, as a useful way to enter a literal
space, instead of invoking <buffy-cycle> or <complete> when typing
space.