convert all dprint() to dprintf()
dprintf() was introduced some years ago but barely used (except in some
other patches).  When Mutt was written it was a C89/C90 project (at
best?) and certain cpp macros and constructs could not be relied upon
(they were GCC extensions).  Now we are C99, and dprintf() allows us
to take advantage of __FILE__, __LINE__, __func__, and __VA_ARGS__ in
useful ways.

We get:
* cleaner, shorter, easier-to-read code
* greater consistency
* all debug messages log the calling function, file, and line,
  without special effort

This patch removes dprint() and mutt_debug() from the code, and
converts every dprint() to a dprintf(). The work was done largely
programmatically (sed, awk, perl) but has been reviewed carefully.
There are a few categories of change here:

* simple conversion, changing syntax and removing the unnecessary `\n`
* duplication of dprintf() where dprint() expected to append conditionally
* pulling up split lines where obviously reasonable
* removal of explicit function names since dprintf() embeds them
* minimal reformatting in service of above goals

Every effort has been made to preserve information and intention.