add deprintf() for including errno
A small number of dprintf() calls explicitly include `errno` as a
parameter. Most of these also include `strerror(errno)`. But the
formatting of these is inconsistent from one instance to another.  This
patch introduces `deprintf()`, which is EXACTLY like `dprintf()` but
includes errno+strerror at the end of the formatted debugging text in a
uniform way across the code.