Add SMTP gsasl auth workaround for broken Microsoft servers.
It appears some Microsoft servers append a non-BASE64 encoded response
to what should be an empty challenge to the initial AUTH sent by the
client.

It's not RFC compliant, so gsasl rightly rejects the response.
However, this doesn't help poor mutt users to authenticate.

To work around this, check the response for the initial reply only,
and if it's not BASE64 decodable, just drop the response.  It is
mostly likely meaningless, and would have resulted in the AUTH failing
regardless.

Thanks to Frank Reker for the bug report and proposed patch, which
this is based upon.