Display the address when failing to add an SMTP recipient
Some providers are picky about recipient addresses, they may refuse to
send mails to an address to which mails have previously been rejected
for "550 5.1.1 Bad destination mailbox address".

Sending an RCPT TO command with such an address will fail with these
providers, and result in a non-standard enhanced status code. In this
case, mutt only emits an error like "SMTP session failed: 550 ..."
without mentioning the address associated with the failed command, which
doesn't help much if there are many recipients.

This patch adds extra error log about the recipient address when an RCPT
TO command fails. Users then could easily distinguish the problematic
recipient, and remove it or contact the provider for support.

Signed-off-by: Yao Zi <me@ziyao.cc>